Rename _roi to just roi and remove methods from top-level
