+++ odc/geo/_overlap.py
@@ -10,7 +10,7 @@ import numpy as np
 from affine import Affine
 from numpy import linalg
 
-from ._roi import roi_boundary, roi_center, roi_from_points, roi_is_empty
+from .roi import roi_boundary, roi_center, roi_from_points, roi_is_empty
 from .geobox import GeoBox, gbox_boundary
 from .math import is_affine_st, maybe_int, snap_scale
