+++ odc/geo/roi.py
